+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Added support for having the default "preemptibility" per top-owner-type read from the default configs configmap in the pod-grouper
 - Added option to profile CPU when running the snapshot tool [#726](https://github.com/NVIDIA/KAI-Scheduler/pull/726) [itsomri](https://github.com/itsomri)
 - GPU resource bookkeeping for DRA enabled resources
+- Add a "tumbling window" usage configuration - calculate a tumbling window size based on a start timne configuration and a duration config field.
 
 ### Fixed
 - Fixed a bug where the snapshot tool would not load topology objects [#720](https://github.com/NVIDIA/KAI-Scheduler/pull/720) [itsomri](https://github.com/itsomri)
@@ -19,6 +20,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - GPU Memory pods are not reclaimed or consolidated correctly
 - Added missing leases permission for the operator [#753](https://github.com/NVIDIA/KAI-Scheduler/pull/753) [dttung2905](https://github.com/dttung2905)
 - Fixed reclaim/preempt/consolidate actions for topology workloads [#739](https://github.com/NVIDIA/KAI-Scheduler/pull/739)  [itsomri](https://github.com/itsomri)
+
+### Changed
+- Renamed the previous "tumbling" option for the scheduler usage window type to "cron".
 
 ## [v0.10.2] - 2025-11-24